--- optuna/storages/_grpc/grpc_imports.py
@@ -1,17 +1,0 @@
-from optuna._imports import try_import
-
-
-with try_import() as _imports:
-    import grpc
-
-    from optuna.storages._grpc.auto_generated import api_pb2
-    from optuna.storages._grpc.auto_generated import api_pb2_grpc
-    from optuna.storages._grpc.auto_generated.api_pb2_grpc import StorageServiceServicer
-
-
-__all__ = [
-    "grpc",
-    "api_pb2",
-    "api_pb2_grpc",
-    "StorageServiceServicer",
-]
